+++ app/workers/netex_full_export_worker.rb
@@ -1,0 +1,13 @@
+class NetexFullExportWorker
+  include Sidekiq::Worker
+  include Concerns::LongRunningWorker
+  extend Concerns::FailingSupport
+
+  # Workers
+  # =======
+
+  def perform(export_id)
+    export = Export::NetexFull.find(export_id)
+    export.run
+  end
+end
